cooking_app/requirements.txt
asgiref==3.6.0
config==0.5.1
decouple==0.0.7
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
itsdangerous==2.1.2
PyJWT==2.6.0
python-decouple==3.7
pytz==2022.7
sqlparse==0.4.3
